+++ core/src/kmem.c
@@ -4,7 +4,7 @@
 /** RK0 - The Embedded Real-Time Kernel '0'                                   */
 /** (C) 2026 Antonio Giacomelli <dev@kernel0.org>                             */
 /**                                                                           */
-/** VERSION: 0.14.1                                                           */
+/** VERSION: 0.14.2                                                           */
 /**                                                                           */
 /** You may obtain a copy of the License at :                                 */
 /** http://www.apache.org/licenses/LICENSE-2.0                                */
@@ -13,6 +13,7 @@
 /******************************************************************************/
 /* COMPONENT: PARTITION MEMORY ALLOCATOR                                      */
 /******************************************************************************/
+
 
 #define RK_SOURCE_CODE